finish project!

+++ app/src/main/java/com/example/screenshotbrainmini/classification/ClassificationResult.kt
@@ -2,6 +2,6 @@ package com.example.screenshotbrainmini.classification
 
 data class ClassificationResult(
     val predictedCategory: String,
-    val confidence: Float,
-    val confidences: Map<String, Float>,
+    val confidence: Double,
+    val confidences: Map<String, Double>,
 )
